feat: get roooms

+++ Models/Room.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Charge_Fee { get; set; }
+        public double Charge_Fee { get; set; }
         public Room(string name, float charge_fee)
         {
             Name = name;
